% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_calcofi_extended}
\alias{area_calcofi_extended}
\title{Area of CalCOFI Extended Stations}
\format{
A single \code{sf} spatial feature
}
\source{
\url{https://calcofi.org/sampling-info/station-positions/}
}
\usage{
area_calcofi_extended
}
\description{
Convex hull of extended CalCOFI stations, additionally clipped by land (\code{rnaturalearthhires::states10}).
}
\concept{data}
\keyword{datasets}
