% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bottle_temp_depth}
\alias{bottle_temp_depth}
\title{Bottle data of temperature with depth (m)}
\format{
A data frame with 1,077 rows and 3 variables \describe{
\item{cast_count}{unique identifier for bottle cast}
\item{depth_m}{depth below the surface in meters}
\item{v}{variable, in this case temperature (Celsius)}
}
}
\source{
\url{https://calcofi.org/sampling-info/station-positions/} \url{https://calcofi.org/data/oceanographic-data/bottle-database/}
}
\usage{
bottle_temp_depth
}
\description{
Extended CalCOFI station bottle cast data with temperature (Celsius) as example data frame
for visualization functions. Data were filtered to casts with a minimum of 50 depth readings.
}
\concept{data}
\keyword{datasets}
