% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bottle_temp_lonlat}
\alias{bottle_temp_lonlat}
\title{Bottle data of temperature in space (latitude, longitude)}
\format{
A data frame with 9,865 rows and 3 variables \describe{
\item{lon}{longitude}
\item{lat}{latitude}
\item{v}{variable, in this case temperature (º Celsius)}
}
}
\source{
\url{https://calcofi.org/sampling-info/station-positions/} \url{https://calcofi.org/data/oceanographic-data/bottle-database/}
}
\usage{
bottle_temp_lonlat
}
\description{
Extended CalCOFI station bottle cast with temperature (º Celsius) as example data frame
for visualization functions.
}
\concept{data}
\keyword{datasets}
