% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcofi4r-package.R
\docType{package}
\name{calcofi4r-package}
\alias{calcofi4r}
\alias{calcofi4r-package}
\title{calcofi4r: CalCOFI R helper functions}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

The California Cooperative Oceanic Fisheries Investigations (CalCOFI) has been surveying the California Current since 1949. This R package provides helper functions to read, analyze and visualize the data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://calcofi.io/calcofi4r}
  \item Report bugs at \url{https://github.com/calcofi/calcofi4r/issues}
}

}
\author{
\strong{Maintainer}: Ben Best \email{ben@ecoquants.com} (\href{https://orcid.org/0000-0002-2686-0784}{ORCID})

Authors:
\itemize{
  \item Ben Best \email{ben@ecoquants.com} (\href{https://orcid.org/0000-0002-2686-0784}{ORCID})
}

Other contributors:
\itemize{
  \item Erin Satterthwaite \email{esatterthwaite@ucsd.edu} (\href{https://orcid.org/0000-0003-0177-7770}{ORCID}) [contributor]
  \item Jennifer Brown \email{jennifer.brown@noaa.gov} [contributor]
  \item Jai Ranganathan \email{jai.ranganathan@gmail.com} [contributor]
  \item Cristina Dobbelaere \email{cdobbelaere@ucsb.edu} [contributor]
  \item CalCOFI [copyright holder, funder]
}

}
\keyword{internal}
