% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_bottle}
\alias{cc_bottle}
\title{Bottle data in space and time}
\format{
A data frame (851,493: rows x columns) with variables:
\describe{
\item{lon}{longitude}
\item{lat}{latitude}
\item{date}{date}
\item{quarter}{quarter}
\item{depth_m}{depth below the surface in meters}
\item{sta_dpos}{difference in position, from 5 (nearshore), 10 (offshore) to 20 (outside 113 station extended area)}
\item{t_degc}{temperature (Celsius)}
\item{salinity}{salinity (TODO: units)}
\item{o2sat}{oxygen saturation (TODO: units)}
}
}
\source{
\url{https://calcofi.org/data/oceanographic-data/bottle-database/}
}
\usage{
cc_bottle
}
\description{
CTD bottle cast data as example data frame for visualization functions.
}
\concept{data}
\keyword{datasets}
