% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{cc_db_connect}
\alias{cc_db_connect}
\title{Connect to the CalCOFI PostgreSQL database (Admin only) - DEPRECATED}
\usage{
cc_db_connect(path_pw = "~/.calcofi_db_pass.txt")
}
\arguments{
\item{path_pw}{path to password file with one line containing the database
password; default: \code{"~/.calcofi_db_pass.txt"}}
}
\value{
a \code{DBI::dbConnect()} object
}
\description{
\lifecycle{deprecated}

This function is deprecated. For the public releases use \code{\link{cc_get_db}}
(DuckDB over Parquet); for the multi-user PostgreSQL database (CTD QA/QC,
database \code{calcofi}) use \code{\link{cc_pg_connect}}, which reads your
role and password from \code{~/.pgpass} instead of a password file.

Note that you must either be running this from the CalCOFI server or have a
\href{https://github.com/calcofi/server#ssh-tunnel-connection-to-postgis-db}{tunnelled SSH connection}
with a user account on the server and the password in a file located at
\code{"~/.calcofi_db_pass.txt"}.
}
\examples{
\dontrun{
con <- cc_db_connect()
DBI::dbListTables(con)
}
}
\concept{database}
