% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_grid_areas}
\alias{cc_grid_areas}
\title{CalCOFI Grid Study Areas}
\format{
A \code{sf} spatial feature set with
\describe{
\item{area_dpos}{character identifier for combinations of difference in position, from 5 (nearshore), 10 (offshore) to 20 (outside 113 station extended area); one of: 5; 10; 20; 5,10; 10,20; 5,10,20}
\item{geom}{dissolved study area from \code{cc_grid} with latitude and longitude in decimal degree geographic coordinates (SRID 4326)}
}
}
\source{
\href{https://calcofi.org/sampling-info/station-positions}{Station Positions – CalCOFI}
}
\usage{
cc_grid_areas
}
\description{
A set of study areas based on dissolving \code{cc_grid} for differentiating
various combinations of 5 (nearshore), 10 (offshore) to 20 (outside 113 station extended area).
For instance, to extract the study area for:
\itemize{
\item only nearshore grid cells, use
\code{dplyr::filter(cc_grid_areas, area_dpos="5")};
\item all grid cells, use
\code{dplyr::filter(cc_grid_areas, area_dpos="5,10,20")};
\item all commonly sampled stations (113 station extended area), use
\code{dplyr::filter(cc_grid_areas, area_dpos="5,10")}.
}
}
\concept{data}
\keyword{datasets}
