% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{cc_list_view_templates}
\alias{cc_list_view_templates}
\title{List Available View Templates}
\usage{
cc_list_view_templates()
}
\value{
Character vector of template names
}
\description{
Returns the names of prebaked view templates that can be used
with \code{\link{cc_make_view}}.
}
\concept{database}
