% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cc_read_bottle}
\alias{cc_read_bottle}
\title{Read CalCOFI bottle data}
\usage{
cc_read_bottle(version = "latest", collect = TRUE, ...)
}
\arguments{
\item{version}{Database version (default: "latest")}

\item{collect}{If TRUE, collect results into memory. If FALSE, return
lazy dbplyr table (default: TRUE)}

\item{...}{Additional filter expressions passed to \code{dplyr::filter()}}
}
\value{
Tibble of bottle data (if collect=TRUE) or lazy table
}
\description{
Convenience function to read bottle sample data from the CalCOFI database.
}
\examples{
\dontrun{
# get all bottle data
bottles <- cc_read_bottle()

# filter by depth
shallow <- cc_read_bottle(depth_m < 100)
}
}
\concept{read}
