% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cc_read_cast}
\alias{cc_read_cast}
\title{Read CalCOFI cast data}
\usage{
cc_read_cast(version = "latest", collect = TRUE, ...)
}
\arguments{
\item{version}{Database version (default: "latest")}

\item{collect}{If TRUE, collect results into memory. If FALSE, return
lazy dbplyr table (default: TRUE)}

\item{...}{Additional filter expressions passed to \code{dplyr::filter()}}
}
\value{
Tibble of cast data (if collect=TRUE) or lazy table
}
\description{
Convenience function to read CTD cast data from the CalCOFI database.
}
\examples{
\dontrun{
casts <- cc_read_cast()
}
}
\concept{read}
