% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cc_read_casts}
\alias{cc_read_casts}
\alias{cc_read_cast}
\title{Read CalCOFI cast data}
\usage{
cc_read_casts(..., version = "latest", collect = TRUE)

cc_read_cast(..., version = "latest", collect = TRUE)
}
\arguments{
\item{...}{Additional filter expressions passed to \code{dplyr::filter()}}

\item{version}{Database version (default: "latest")}

\item{collect}{If TRUE, collect results into memory. If FALSE, return
lazy dbplyr table (default: TRUE)}
}
\value{
Tibble of casts data (if collect=TRUE) or lazy table
}
\description{
Convenience function to read CTD/bottle cast data from the CalCOFI database.
The data is stored in the \code{casts} table.
}
\examples{
\dontrun{
# get first 100 casts
casts <- cc_read_casts() |> head(100)
}
}
\concept{read}
