% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cc_read_cruise}
\alias{cc_read_cruise}
\title{Read CalCOFI cruise data}
\usage{
cc_read_cruise(..., version = "latest", collect = TRUE)
}
\arguments{
\item{...}{Additional filter expressions passed to \code{dplyr::filter()}}

\item{version}{Database version (default: "latest")}

\item{collect}{If TRUE, collect results into memory. If FALSE, return
lazy dbplyr table (default: TRUE)}
}
\value{
Tibble of cruise data (if collect=TRUE) or lazy table
}
\description{
Convenience function to read cruise metadata from the CalCOFI database.
}
\examples{
\dontrun{
cruises <- cc_read_cruise()
}
}
\concept{read}
