% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cc_read_larvae}
\alias{cc_read_larvae}
\title{Read CalCOFI larvae data}
\usage{
cc_read_larvae(version = "latest", collect = TRUE, ...)
}
\arguments{
\item{version}{Database version (default: "latest")}

\item{collect}{If TRUE, collect results into memory. If FALSE, return
lazy dbplyr table (default: TRUE)}

\item{...}{Additional filter expressions passed to \code{dplyr::filter()}}
}
\value{
Tibble of larvae data (if collect=TRUE) or lazy table
}
\description{
Convenience function to read larval fish data from the CalCOFI database.
}
\examples{
\dontrun{
# get all larvae data
larvae <- cc_read_larvae()

# get specific species
engraulis <- cc_read_larvae(species_id == 123)
}
}
\concept{read}
