% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cc_read_site}
\alias{cc_read_site}
\title{Read CalCOFI site data}
\usage{
cc_read_site(..., version = "latest", collect = TRUE)
}
\arguments{
\item{...}{Additional filter expressions passed to \code{dplyr::filter()}}

\item{version}{Database version (default: "latest")}

\item{collect}{If TRUE, collect results into memory. If FALSE, return
lazy dbplyr table (default: TRUE)}
}
\value{
Tibble of site data (if collect=TRUE) or lazy table
}
\description{
Convenience function to read sampling site data from the CalCOFI database.
}
\examples{
\dontrun{
sites <- cc_read_site()
}
}
\concept{read}
