% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cc_read_tow}
\alias{cc_read_tow}
\title{Read CalCOFI tow data}
\usage{
cc_read_tow(..., version = "latest", collect = TRUE)
}
\arguments{
\item{...}{Additional filter expressions passed to \code{dplyr::filter()}}

\item{version}{Database version (default: "latest")}

\item{collect}{If TRUE, collect results into memory. If FALSE, return
lazy dbplyr table (default: TRUE)}
}
\value{
Tibble of tow data (if collect=TRUE) or lazy table
}
\description{
Convenience function to read net tow data from the CalCOFI database.
}
\examples{
\dontrun{
tows <- cc_read_tow()
}
}
\concept{read}
