% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_index}
\alias{create_index}
\title{Create index in database}
\usage{
create_index(
  con,
  tbl,
  flds,
  is_geom = F,
  is_unique = F,
  overwrite = F,
  show = F,
  exec = T
)
}
\arguments{
\item{con}{database connection object from \code{DBI::dbConnect()}, e.g. from \code{cc_db_connect()}}

\item{tbl}{table name as character}

\item{flds}{character vector of fields in table used for index}

\item{is_geom}{logical (default: FALSE) whether geometry field, so create GIST() index}

\item{is_unique}{logical (default: FALSE) whether to impose a unique constraint, to prevent duplicates; default: FALSE}

\item{overwrite}{logical (default: FALSE) whether to overwrite existing index}

\item{show}{logical (default: FALSE) whether to show SQL statement}

\item{exec}{logical (default: TRUE) whether to execute SQL statement}
}
\value{
nothing
}
\description{
Create index in database
}
\examples{
\dontrun{
con <- cc_db_connect()
create_index(con, "ctd_casts", "geom", is_geom=T)
create_index(con, "ctd_casts", "cast_count", is_unique=T)
}
}
\concept{database}
