% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{get_oceano_var_aoi}
\alias{get_oceano_var_aoi}
\title{Get oceanographic variable for area of interest}
\usage{
get_oceano_var_aoi(
  var,
  aoi,
  date_step = c("year", "day", "week", "month", "quarter", "decade"),
  depth_min = 0,
  depth_max = 10
)
}
\arguments{
\item{var}{variable of interest (TODO: see keys)}

\item{aoi}{area of interest}

\item{date_step}{eg month, quarter, year}

\item{depth_min}{depth minimum}

\item{depth_max}{depth maximum}
}
\value{
data frame of values
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated because it relies on legacy data structures
(\code{var_lookup}) that are not compatible with the new DuckDB database.
Query bottle measurement data directly with \code{\link[=cc_get_db]{cc_get_db()}} and
\code{\link[=cc_read_measurements]{cc_read_measurements()}}, then use sf for spatial filtering.
}
\examples{
\dontrun{
# deprecated - use DuckDB queries instead:
con <- cc_get_db()
d <- DBI::dbGetQuery(con, "
  SELECT s.longitude, s.latitude, o.datetime, o.depth_min_m,
         o.measurement_value, o.measurement_type
  FROM obs_env o
  JOIN sample s USING (sample_key)
  WHERE o.dataset_key = 'calcofi_bottle' AND o.measurement_type = 'oxygen_ml_l'")
# then filter spatially with sf::st_filter(d_sf, aoi)
}
}
\concept{analyze}
