% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{map_contours}
\alias{map_contours}
\title{Map contours from station data using a GAM}
\usage{
map_contours(df, ply, k = 60, cw = 0.1)
}
\arguments{
\item{df}{data frame containing the fields: \code{lon} (longitude), \code{lat}
(latitude) and \code{v} (value)}

\item{ply}{polygon describing the study area mask}

\item{k}{cell width in decimal degrees to run interpolation; default: \code{0.1}}
}
\value{
polygon simple features (\code{sf}) of contour \code{isoband::isobands()}
}
\description{
Fit generalized additive model to a variable in space using a smoother on
latitude & longitude. Further clip by bounding polygon.
}
\examples{
v_ply <- map_contours(bottle_temp_lonlat, area_calcofi_extended)
mapview::mapView(v_ply, zcol="v", layer.name="temp(ºC)")
}
\concept{visualize}
