% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{map_raster}
\alias{map_raster}
\title{Map raster}
\usage{
map_raster(r, legend_title = "Temperature (ºC)")
}
\arguments{
\item{r}{raster of type \code{raster::raster()}}

\item{legend_title}{title for legend of variable mapped}
}
\value{
interactive plot of \code{leaflet::leaflet()}
}
\description{
Map raster
}
\examples{
# get variables
(v <- get_variables())

# get data for the first variable
(d <- get_timeseries(v$table_field[1]))

# plot time series with the first variable
with(v[1,],
  plot_timeseries(
    # data and columns (from d)
    d, year, t_deg_c_avg, t_deg_c_sd,
    # plot attributes (from v)
    plot_title, plot_label, plot_color))
}
\concept{visualize}
