% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{modal_depth_profile}
\alias{modal_depth_profile}
\title{Depth Profile Modal Dialog}
\usage{
modal_depth_profile(map_sp)
}
\arguments{
\item{map_sp}{maplibre object (currently unused in implementation, retained for future enhancement)}
}
\value{
Shiny modal dialog object with transect drawing interface and buffer distance input
}
\description{
Creates a modal dialog for defining a transect line segment and buffer
distance to generate environmental depth profiles.
}
\details{
Users draw a line segment on the map to define a transect. The buffer distance
controls the width of the corridor around the transect for data aggregation.
Default buffer is 5 km.
}
\examples{
\dontrun{
# in server.R
observeEvent(input$create_profile, {
  showModal(modal_depth_profile(map_sp = NULL))
})
}

}
\seealso{
\code{\link{buffer_transect}} for transect buffer generation
}
