% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{pts_to_rast_idw}
\alias{pts_to_rast_idw}
\title{Interpolate points to raster using Inverse-Distance Weighting (IDW)}
\usage{
pts_to_rast_idw(
  pts,
  fld,
  aoi,
  radius = 1.5,
  power = 1.3,
  smooth = 0.2,
  maxPoints = Inf,
  minPoints = 1,
  near = FALSE,
  fill = NA,
  verbose = T
)
}
\arguments{
\item{pts}{input points as simple features \code{sf::sf}}

\item{fld}{quoted string of field from \code{pts} with z-value to interpolate}

\item{aoi}{area of interest as simple features \code{sf::sf}}

\item{radius}{argument for \code{terra::interpIDW()}; default = \code{1.5}}

\item{power}{argument for \code{terra::interpIDW()}; default = \code{1.3}}

\item{smooth}{argument for \code{terra::interpIDW()}; default = \code{0.2}}

\item{maxPoints}{argument for \code{terra::interpIDW()}; default = \code{Inf}}

\item{minPoints}{argument for \code{terra::interpIDW()}; default = \code{1}}

\item{near}{argument for \code{terra::interpIDW()}; default = \code{FALSE}}

\item{fill}{argument for \code{terra::interpIDW()}; default = \code{NA}}

\item{verbose}{show messages of all variables used; default = FALSE}
}
\value{
terra::rast() object
}
\description{
Interpolate from points to raster using inverse-distance weighting
(\code{terra::interpIDW()}). The further the input point from the destination pixel
in the output raster, the less its value is applied. The dimensions of the
output raster are automatically calculated based on approaching 100 pixels in
the maximum dimension (of longitude or latitude) based on the extent of the
area of interest.
}
\concept{analyze}
