% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{rast_to_contours}
\alias{rast_to_contours}
\title{Raster to contour polygons}
\usage{
rast_to_contours(r, aoi, n_brks = 8)
}
\arguments{
\item{r}{input raster as \code{terra::rast}}

\item{aoi}{area of interest as simple features \code{sf::sf}}

\item{n_brks}{number of breaks}
}
\value{
contour polygon simple features (\code{sf::sf})
}
\description{
Contour the input raster (using \code{stars::st_contour}). Use "pretty" breaks
(using \code{classInt::classIntervals()}). Clip by area of interest.
}
\concept{analyze}
