% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Oceanographic stations}
\format{
A data frame with 2634 rows and 11 variables.
\describe{
\item{Sta_ID}{Line and Station}
\item{Sta_ID_line}{Line data - the first half of the value from the Sta_ID field}
\item{Sta_ID_station}{Station data - the second half of the value from the Sta_ID field}
\item{lon}{Station longitude}
\item{lat}{Station latitude}
\item{offshore}{A Boolean variable, with "TRUE" indicating an offshore CalCOFI station.}
\item{is_cast}{To be added}
\item{is_cce}{To be added}
\item{is_ccelter}{To be added}
\item{is_sccoos}{To be added}
\item{geometry}{A vector containing the station latitude and longitude}
}
}
\source{
\url{https://calcofi.org/data/oceanographic-data/bottle-database/}
}
\usage{
stations
}
\description{
The geographic locations of every bottle sampling station utilized on a CalCOFI
cruise. This data set is an extraction and modification of the CalCOFI cast table.
}
\concept{data}
\keyword{datasets}
