% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erd.R
\name{updateMermaid}
\alias{updateMermaid}
\title{Update the mermaid library in the package}
\usage{
updateMermaid(version = "")
}
\arguments{
\item{version}{A \link{character} of the desired version (the latest by default)}
}
\value{
See value returned by \link{download.file}
}
\description{
Update the mermaid library in the package
}
