% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{CC_AREAL_GEARS}
\alias{CC_AREAL_GEARS}
\title{Gears whose \code{std_haul_factor} standardizes to a depth-integrated 10 m2 (oblique / vertical tows)}
\usage{
CC_AREAL_GEARS
}
\description{
Gears whose \code{std_haul_factor} standardizes to a depth-integrated 10 m2 (oblique / vertical tows)
}
\concept{quality}
