% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{cc_db_catalog}
\alias{cc_db_catalog}
\title{Show CalCOFI database catalog as interactive table}
\usage{
cc_db_catalog(tables = NULL)
}
\arguments{
\item{tables}{optional character vector of table names to filter; default: \code{NULL}}
}
\value{
DT::datatable() of tables and columns
}
\description{
Read the tables and columns from the following \href{https://api.calcofi.io}{CalCOFI API}
and display as an interactive table (as \code{DT::datatable()}):
\itemize{
\item \href{https://api.calcofi.io/db_tables}{api.calcofi.io/db_tables}
\item \href{https://api.calcofi.io/db_columns}{api.calcofi.io/db_columns}
}
}
\examples{
# full catalog
cc_db_catalog()
# only certain tables
cc_db_catalog(tables = c("larvae_counts","nets","tows","stations","cruises"))
}
\concept{database}
