% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\docType{data}
\name{CC_DENSITY_UNITS}
\alias{CC_DENSITY_UNITS}
\title{Units the release publishes as densities (per unit area, per unit volume)}
\format{
An object of class \code{list} of length 2.
}
\usage{
CC_DENSITY_UNITS
}
\description{
Units the release publishes as densities (per unit area, per unit volume)
}
\concept{quality}
\keyword{datasets}
