% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_grid}
\alias{cc_grid}
\title{CalCOFI Grid for Extracting Effort}
\format{
A \code{sf} spatial feature set with
\describe{
\item{sta_key}{station key in the form of "\code{lin},\code{pos}"}
\item{sta_lin}{alongshore line in CalCOFI coordinate system}
\item{sta_pos}{offshore position in CalCOFI coordinate system}
\item{sta_dpos}{difference in position, from 5 (nearshore), 10 (offshore) to 20 (outside 113 station extended area)}
\item{sta_pattern}{the CalCOFI station pattern; one of: "standard", "extended" or "historical"}
\item{sta_shore}{the position wrt shore; one of: "nearshore" or "offshore"}
\item{geom}{station voronoi polygon with latitude and longitude in decimal degree geographic coordinates (SRID 4326)}
}
}
\source{
\href{https://calcofi.org/sampling-info/station-positions}{Station Positions – CalCOFI}
}
\usage{
cc_grid
}
\description{
A grid for calculating effort by station using Voronoi diagram to fetch nearest station, additionally clipped by land (\code{rnaturalearthhires::states10}).
}
\concept{data}
\keyword{datasets}
