% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_grid_ctrs}
\alias{cc_grid_ctrs}
\title{CalCOFI Grid Centroids for Extracting Effort}
\format{
A \code{sf} spatial feature set with
\describe{
\item{sta_key}{station key in the form of "\code{lin},\code{pos}"}
\item{sta_lin}{alongshore line in CalCOFI coordinate system}
\item{sta_pos}{offshore position in CalCOFI coordinate system}
\item{sta_dpos}{difference in position, from 5 (nearshore), 10 (offshore) to 20 (outside 113 station extended area)}
\item{geom}{station latitude and longitude in decimal degree geographic coordinates (SRID 4326)}
}
}
\source{
\url{https://calcofi.org/sampling-info/station-positions/}
}
\usage{
cc_grid_ctrs
}
\description{
A set of centroids for the grid to calculate effort by station using Voronoi diagram to fetch nearest station, additionally clipped by land (\code{rnaturalearthhires::states10}).
}
\concept{data}
\keyword{datasets}
