% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_grid_zones}
\alias{cc_grid_zones}
\title{CalCOFI Grid Zones}
\format{
A \code{sf} spatial feature set with 6 rows × 9 columns:
\describe{
\item{zone_key}{unique zone key of the form \code{"{sta_pattern}-{sta_shore}"} }
\item{sta_pattern}{the CalCOFI station pattern; one of: "standard", "extended" or "historical"}
\item{sta_shore}{the position wrt shore; one of: "nearshore" or "offshore"}
\item{sta_dpos}{the difference in position: 5 (nearshore), 10 (offshore) or 20 (historical)}
\item{sta_lin_min}{the minimum value of dissolved \code{sta_lin} from \code{cc_grid}}
\item{sta_lin_max}{the maximum value of dissolved \code{sta_lin} from \code{cc_grid}}
\item{sta_pos_min}{the minimum value of dissolved \code{sta_pos} from \code{cc_grid}}
\item{sta_pos_max}{the maximum value of dissolved \code{sta_pos} from \code{cc_grid}}
\item{geom}{geometry of dissolved zone from \code{cc_grid} with latitude and longitude in decimal degree geographic coordinates (EPSG:4326)}
}
}
\source{
\href{https://calcofi.org/sampling-info/station-positions}{Station Positions – CalCOFI}
}
\usage{
cc_grid_zones
}
\description{
A set of zones based on dissolving \code{cc_grid} for differentiating position wrt
the shore (\code{sta_shore}: "nearshore" or "offshore") and station patterns (
\code{sta_pattern}: "standard", "extended" or "historical").
}
\concept{data}
\keyword{datasets}
