% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cc_list_tables}
\alias{cc_list_tables}
\title{List tables in CalCOFI database}
\usage{
cc_list_tables(version = "latest")
}
\arguments{
\item{version}{Database version (default: "latest")}
}
\value{
Character vector of table names
}
\description{
Lists all available tables in a CalCOFI database release.
}
\examples{
\dontrun{
cc_list_tables()
}
}
\concept{read}
