% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_places}
\alias{cc_places}
\title{CalCOFI Places}
\format{
A \code{sf} spatial feature set with
\describe{
\item{key}{character key uniquely identifying the record}
\item{category}{character key}
\item{name}{name of the place, given the category}
\item{geom}{polygon geometry in geographic coordinates (SRID 4326)}
}
}
\usage{
cc_places
}
\description{
A set of places for commonly extracting CalCOFI data.
}
\details{
There are three categories of places [key]:
\enumerate{
\item CalCOFI
\itemize{
\item Standard Nearshore [cc_extended-nearshore]
\item Standard Offshore [cc_extended-offshore]
\item Extended Nearshore [cc_historical-nearshore]
\item Extended Offshore [cc_historical-offshore]
\item Historical Nearshore [cc_standard-nearshore]
\item Historical Offshore [cc_standard-offshore]
}
\item Integrated Ecosystem Assessment
\itemize{
\item California Current [iea_ca]
}
\item National Marine Sanctuary
\itemize{
\item Cordell Bank [nms_cb]
\item Channel Islands [nms_ci]
\item Greater Farallones [nms_gf]
\item Monterey Bay [nms_mb]
\item Olympic Coast [nms_oc]
}
}
}
\concept{data}
\keyword{datasets}
