% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cc_read_species}
\alias{cc_read_species}
\title{Read CalCOFI species data}
\usage{
cc_read_species(..., version = "latest", collect = TRUE)
}
\arguments{
\item{...}{Additional filter expressions passed to \code{dplyr::filter()}}

\item{version}{Database version (default: "latest")}

\item{collect}{If TRUE, collect results into memory. If FALSE, return
lazy dbplyr table (default: TRUE)}
}
\value{
Tibble of species data (if collect=TRUE) or lazy table
}
\description{
Convenience function to read species taxonomy data from the CalCOFI database.
}
\examples{
\dontrun{
species <- cc_read_species()
}
}
\concept{read}
