% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dic}
\alias{dic}
\title{Oceanographic DIC data}
\format{
TBD...e.g., A data frame with 53940 rows and 10 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
dic
}
\description{
TBD...e.g., A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\concept{data}
\keyword{datasets}
