% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_cruises}
\alias{get_cruises}
\title{Get cruises from CalCOFI API}
\usage{
get_cruises()
}
\value{
data frame with columns:
\itemize{
\item \code{cruise_id}: unique identifier for cruise
\item \code{date_beg}: date cruise began
\item \code{date_end}: date cruise ended
\item \code{lon_min}: longitude, minimum (as in for bounding box of cruise)
\item \code{lon_max}: longitude, minimum (as in for bounding box of cruise)
\item \code{lat_min}: latitude, minimum (as in for bounding box of cruise)
\item \code{lat_max}: latitude, minimum (as in for bounding box of cruise)
\item \code{n_casts}: number of oceanographic CTD casts made for given cruise
}
}
\description{
Get cruises from CalCOFI API
}
\examples{
get_cruises()
}
\concept{read}
