% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{get_oceano_var_aoi}
\alias{get_oceano_var_aoi}
\title{Get oceanographic variable for area of interest}
\usage{
get_oceano_var_aoi(
  var,
  aoi,
  date_step = c("year", "day", "week", "month", "quarter", "decade"),
  depth_min = 0,
  depth_max = 10
)
}
\arguments{
\item{var}{variable of interest (TODO: see keys)}

\item{aoi}{area of interest}

\item{date_step}{eg month, quarter, year}

\item{depth_min}{depth minimum}

\item{depth_max}{depth maximum}
}
\value{

}
\description{
Get oceanographic variable for area of interest
}
\examples{
\dontrun{
get_oceano_var_aoi("Bottle O2(ml_L)", cinms_ply, "year", 0, 4000)
}

}
\concept{analyze}
