% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_raster}
\alias{get_raster}
\title{Get raster of interpolated values from CalCOFI API}
\usage{
get_raster(
  variable = "ctdcast_bottle.t_deg_c",
  cruise_id = "1949-03-01-C-31CR",
  depth_m_min = NULL,
  depth_m_max = NULL,
  out_tif
)
}
\arguments{
\item{variable}{Variable to fetch from the CalCOFI API. One of \code{table_field} values from \code{get_variables()}.
Default is \code{"ctdcast_bottle.t_deg_c"}.}

\item{cruise_id}{Cruise identifier. One of \code{cruise_id} values from \code{get_cruises()}.
Default is the first cruise "1949-03-01-C-31CR".}

\item{depth_m_min}{Minimum depth range in meters, e.g. 0. Default is \code{NULL}, as in not filtered.}

\item{depth_m_max}{Maximum depth range in meters, e.g. 5351. Default is \code{NULL}, as in not filtered.}

\item{out_tif}{Output path to write raster (*.tif)}
}
\value{
path to output raster in GeoTIFF format (*.tif)
}
\description{
Get raster of interpolated values from CalCOFI API
}
\examples{
out_tif <- tempfile(fileext=".tif")
get_raster(
  variable = "ctdcast_bottle.t_deg_c",
  cruise_id = "1949-03-01-C-3d1CR",
  depth_m_min = 0, depth_m_max = 200,
  out_tif = out_tif)
r <- raster::raster(out_tif)
raster::plot(r)
unlink(out_tif)
}
\concept{read}
