% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Get timeseries summary from CalCOFI API}
\usage{
get_timeseries(
  variable = "ctdcast_bottle.t_deg_c",
  aoi_wkt = NULL,
  depth_m_min = NULL,
  depth_m_max = NULL,
  date_beg = NULL,
  date_end = NULL,
  time_step = "year",
  stats = c("p10", "mean", "p90")
)
}
\arguments{
\item{variable}{Variable to fetch from the CalCOFI API. One of \code{table_field} values from \code{get_variables()}.
Default is \code{"ctdcast_bottle.t_deg_c"}.}

\item{aoi_wkt}{Area of interest (AOI), spatially described as
\href{https://en.wikipedia.org/wiki/Well-known_text_representation_of_geometry}{well-known text (WKT)}.}

\item{depth_m_min}{Minimum depth range in meters, e.g. 0. Default is \code{NULL}, as in not filtered.}

\item{depth_m_max}{Maximum depth range in meters, e.g. 5351. Default is \code{NULL}, as in not filtered.}

\item{date_beg}{Beginning of date range, e.g."1949-02-28". Default is \code{NULL}, as in not filtered.}

\item{date_end}{End of date range, e.g. "2020-01-26". Default is \code{NULL}, as in not filtered.}

\item{time_step}{Time step over which to summarize. One of: a sequential
increment ("decade","year","year.quarter","year.month","year.week","date"),
or a climatology ("quarter","month","week","julianday","hour").
Default is \code{"year"}.}

\item{stats}{Statistics to show per \code{date_step}. Acceptable values include any combination of:
"avg", "median", "min", "max", "sd" or "p#"
where "sd" is the standard deviation and
"p#" represents the percentile value 0 to 100 within available range of values.
Default is \code{c("p10", "mean", "p90")}.}
}
\description{
Get timeseries summary from CalCOFI API
}
\examples{
get_timeseries(
  variable = "ctdcast_bottle.t_deg_c",
  aoi_wkt = "POLYGON ((-120.6421 33.36241, -118.9071 33.36241, -118.9071 34.20707, -120.6421 34.20707, -120.6421 33.36241))",
  depth_m_min = 0, depth_m_max = 200,
  date_beg = "2000-01-01", date_end = "2020-01-01",
  time_step = "year",
  stats = c("p10", "mean", "p90"))

}
\concept{read}
