% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_variables}
\alias{get_variables}
\title{Get variables from CalCOFI API}
\usage{
get_variables()
}
\value{
data frame with columns:
\itemize{
\item \code{category}: category, e.g. "Oceanographic"
\item \code{table_field}: unique identifier of the variable given by the form of table.field
\item \code{plot_title}: title used in a typical time series plot output
\item \code{plot_label}: y-axis label used in a typical time series plot output
}
}
\description{
Get variables from CalCOFI API
}
\concept{read}
