% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{map_contours}
\alias{map_contours}
\title{Map contours from station data using a GAM}
\usage{
map_contours(df, ply, gam_k = 60, grid_width = 0.1, n_breaks = 7)
}
\arguments{
\item{df}{data frame containing the fields: \code{lon} (longitude), \code{lat}
(latitude) and \code{v} (value)}

\item{ply}{polygon describing the study area mask}

\item{gam_k}{number of knots to use on the spatial smoother \code{s(lon, lat)};
default: \code{100}}

\item{grid_width}{cell width in decimal degrees to run interpolation; default: \code{0.1}}

\item{n_breaks}{number of breaks in values for creating contours; default: \code{7}}
}
\value{
polygon simple features (\code{sf}) of contour \code{isoband::isobands()}
}
\description{
Fit generalized additive model to a variable in space using a smoother on
latitude & longitude. Further clip by bounding polygon.
}
\examples{
v_ply <- map_contours(bottle_temp_lonlat, area_calcofi_extended)
mapview::mapView(v_ply, zcol="v", layer.name="temp(ºC)")
}
\concept{visualize}
