% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{map_contours}
\alias{map_contours}
\title{Table to Contour Map}
\usage{
map_contours(df, ply, k = 60, cw = 0.1)
}
\arguments{
\item{df}{data frame containing the fields: \code{lon} (longitude), \code{lat}
(latitude) and \code{v} (value)}

\item{ply}{polygon describing the study area mask}

\item{k}{cell width in decimal degrees to run interpolation; default: \code{0.1}}
}
\description{
Fit generalized additive model to a variable in space using a smoother on
latitude & longitude. Further clip by bouding polygon.
}
\examples{
v_ply <- map_contours(stations_t_degc, area_calcofi_extended)
mapview::mapView(v_ply, zcol="v", layer.name="temp(ºC)")
}
\concept{visualize}
