% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{map_raster}
\alias{map_raster}
\title{Map raster interactively}
\usage{
map_raster(r, legend_title = "Temperature (C)")
}
\arguments{
\item{r}{raster of type \code{raster::raster()}}

\item{legend_title}{title for legend of variable mapped}
}
\value{
interactive plot of \code{leaflet::leaflet()}
}
\description{
Map raster of interpolated oceanographic variable for a cruise.
}
\examples{
(r_tif <- tempfile(fileext=".tif"))

# use second variable from previously fetched v
c(v$table_field[2], v$plot_label[2])

# fetch interpolated raster from CalCOFI API
get_raster(
  variable = v$table_field[2],
  cruise_id = "2020-01-05-C-33RL",
  depth_m_min = 0, depth_m_max = 200,
  out_tif = r_tif)

r <- raster::raster(r_tif)

# plot raster
map_raster(r, v$plot_label[2])
}
\concept{visualize}
