% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot timeseries}
\usage{
plot_timeseries(d)
}
\arguments{
\item{d}{data frame with expected fields (\code{Date}, \code{var_q10}, \code{var_avg}, \code{var_q90}) from \code{\link[=get_oceano_var_aoi]{get_oceano_var_aoi()}}}
}
\value{
\link[dygraphs:dygraph]{dygraphs::dygraph}
}
\description{
Plot timeseries
}
\concept{visualize}
